[optimize] UserProfilePage add follow button
